refactor: 사용자 정보 조회 grade, param 변경
