+++ src/main/java/com/itplace/userapi/user/service/UserService.java
@@ -1,14 +1,14 @@
 package com.itplace.userapi.user.service;
 
 import com.itplace.userapi.security.verification.email.dto.EmailConfirmRequest;
-import com.itplace.userapi.user.dto.UserInfoDto;
 import com.itplace.userapi.user.dto.request.FindEmailConfirmRequest;
 import com.itplace.userapi.user.dto.request.ResetPasswordRequest;
 import com.itplace.userapi.user.dto.response.FindEmailResponse;
 import com.itplace.userapi.user.dto.response.FindPasswordConfirmResponse;
+import com.itplace.userapi.user.dto.response.UserInfoResponse;
 
 public interface UserService {
-    UserInfoDto getUserInfo(Long userId);
+    UserInfoResponse getUserInfo(Long userId);
 
     FindEmailResponse findEmailConfirm(FindEmailConfirmRequest request);
